requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
bitsandbytes==0.48.1
braceexpand==0.1.7
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.3.0
dill==0.4.0
distro==1.9.0
faiss==1.9.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
higher==0.2.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
iopath==0.1.10
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.14.0
opencv-python==4.12.0.88
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
progressbar2==4.5.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
segment-anything==1.0
sentence-transformers==5.1.1
sentencepiece==0.2.1
sentry-sdk==2.42.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.3
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
visual-genome==1.1.1
wandb==0.22.2
webdataset==1.0.2
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
